Fixed some errors. First Commit. See details.
Still not a working branch but very Close to it.
I changed the compiler of the SDK to GCC 4.2 and compiler for the
xCodeProject to GCC 4.2 as well.
It compiles very nice but the xThreadMain is not found in the code.
Have to find a solution for this.
This is also a first commit of the mac Port project. Open with xCode
3.2.6.

VERY IMPORTANT:
Only compiles under 32Bit Target. 64Bit is given us an error with an
Array definition.