More debug and clean up
Cleaned up a part of the code. Added a lot of debug code stuff. Tried to
fix the normal problem. It seems the engine does not care even if i do
not add normals the mesh looks the same as when i am adding my vertex
normals. Project is also easier and cleaner to build now. Just press
build and it should run.