Fixed broken csproj files and therefore the solution
When initializing a new .net core or standard project in the same solution with VS 2019 and the built-in assistant, and there is already such a project present, everything within these two projects crashes and refuses to build. The solution is to remove the <Project SDK=.net...> line and replace it with <Project> and import the .net SDK targets in the file.